Extract review report module
